requirements.txt
asgiref==3.8.1
Django==5.0.4
django-rest==0.8.7
djangorestframework==3.15.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
